rename test util
